social-brain-ai-service/SocialBrain/image_generation.py
import os
from dotenv import load_dotenv

# Try to import the optional image generation client
try:
    import google.generativeai as genai
except ImportError:
    genai = None

# Load your API key
load_dotenv()

# Configure Google Generative AI if available
if genai is not None:
    genai.configure(api_key=os.getenv("GOOGLE_API_KEY"))

def generate_image(prompt: str, size: str = "1024x1024", quality: str = "standard", n: int = 1):
    """
    Generate image using Google's Imagen API.
    Note: Google Gemini doesn't have built-in image generation like DALL-E.
    This function uses Google's Generative AI for image generation.
    You may need to use Google Cloud Imagen API or another service.
    
    For now, this returns a placeholder. You can:
    1. Use Google Cloud Imagen API (requires GCP setup)
    2. Keep using OpenAI DALL-E for images
    3. Use another image generation service
    """
    # If the image client isn't installed, just log and return placeholders
    if genai is None:
        print("Image generation requested but 'google.generativeai' is not installed.")
        print("Skipping image generation and returning placeholders.")
        return None, None

    try:
        # Placeholder for actual image generation implementation.
        print(f"Image generation requested for: {prompt}")
        print("Note: Google Gemini doesn't include image generation. Consider using Google Cloud Imagen API or keeping DALL-E for images.")

        # TODO: Implement real image generation via Google Cloud Imagen or another service.
        return None, None
    except Exception as e:
        print(f"Error generating image: {e}")
        return None, None